started adding scatter dynamic modifiers to descriptors and SU

+++ spike/riscv/insns/so_p_one.h
@@ -7,7 +7,7 @@ auto destElements = destPReg.getPredicate();
 
 std::vector<uint8_t> predicate(destPReg.vLen, 1);
 
-for (int i = 0; i < destPReg.vLen; i++){
+for (size_t i = 0; i < destPReg.vLen; i++){
     if (!pi.at(i))
         predicate.at(i) = destElements.at(i); // merging
 }
